01_simple_agent/pyproject.toml
[project]
name = "adk-worskpace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.79.0",
    "google-adk>=1.25.0",
    "google-cloud-secret-manager>=2.26.0",
]
